+++ lib/Models/WebMapServiceCatalogItem.ts
@@ -1149,6 +1149,8 @@ class WebMapServiceCatalogItem
           ...dimensionParameters,
           styles: this.styles === undefined ? "" : this.styles
         },
+        tileWidth: this.tileWidth,
+        tileHeight: this.tileHeight,
         tilingScheme: /*defined(this.tilingScheme) ? this.tilingScheme :*/ new WebMercatorTilingScheme(),
         maximumLevel: maximumLevel,
         rectangle: rectangle,